+++ src/main/java/org/example/petcareplus/dto/PostDTO.java
@@ -2,7 +2,7 @@ package org.example.petcareplus.dto;
 import org.example.petcareplus.entity.Post;
 
 public class PostDTO {
-    private Integer postId;
+    private Long postId;
     private String title;
     private String description;
     private Integer rating;
@@ -19,8 +19,8 @@ public class PostDTO {
     }
 
     // Getter + Setter
-    public Integer getPostId() { return postId; }
-    public void setPostId(Integer postId) { this.postId = postId; }
+    public Long getPostId() { return postId; }
+    public void setPostId(Long postId) { this.postId = postId; }
 
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
